refactor: 랜덤 온도 입력 관련 폴더 구조
# 폴더 구조 변경

- 페이지 단위 (pages)
  1. RandomMachineTemp, RandomFoodTemp: 랜덤 온도 입력 페이지는 페이지 단위로 pages 폴더로 이동합니다.

- 위젯 단위 (widgets)
  1. RandomTempForm: 랜덤으로 온도를 입력하거나, 설정한 온도를 저장하는 기능을 수행하는 UI 요소로 특정 페이지(기기 및 식품 온도 랜덤 제출 페이지)에서만 사용되는 컴포넌트로 widgets 폴더로 이동합니다.

- 기능 단위 (features)
  1. useGenerateRandomTemp: 사용자가 설정한 최소, 최대 온도 사이의 랜덤한 온도를 생성해 새로운 배열을 리턴하는 기능을 제공하는 함수로 features 폴더로 이동합니다.
  2. RandomTemp: rc-slider의 슬라이더 기능을 이용해 사용자에게 온도 조절 기능을 제공하는 컴포넌트로 features 폴더로 이동합니다.

- 엔티티 단위 (entities)
  1. RandomTemp: 기존의 data-mutation 및 비즈니스 로직을 담당하던 커스텀 훅을 분리해 entities 폴더로 이동합니다. dataMutation 함수는 query 파일, 온도 저장과 제출 함수는 action 파일로 관리합니다.

- 공통 단위 (shared)
  1. useModalWithTemp: 제출과 관련된 4개의 페이지에서 모달의 생성 유무를 result 값을 통해 확인하기 때문에 해당 기능을 위한 커스텀 훅을 생성했습니다. 이 때 handleRetry 함수로 제출 실패시 사용하는 submit 함수를 인자로 전달해 사용합니다.

- 제출 관련 폴더를 전부 수정함에 따라 기존의 Modal 컴포넌트는 삭제했습니다.
