--- src/components/Modal.module.css
@@ -1,4 +1,0 @@
-.section {
-  display: flex;
-  font-family: "Pretendard";
-}
