+++ src/entities/custom-products/input-temp/model/query.js
@@ -4,7 +4,7 @@ import {
   getCustomMachines,
   submitFoods,
   submitMachines,
-} from "../../../api/Products";
+} from "../../../../api/Products";
 
 function dataMutation(queryKey, getProductFunc, submitFunc, productKey) {
   const productsQuery = useQuery([queryKey], getProductFunc, {
