+++ src/entities/custom-products/random-temp/index.js
@@ -1,0 +1,4 @@
+export { useRandomFoodTemp } from "./model/query";
+export { useRandomMachineTemp } from "./model/query";
+export { useSubmitRandomRange } from "./model/action";
+export { useSaveRandomRange } from "./model/action";
