+++ src/features/generate-random-temp/index.js
@@ -1,0 +1,1 @@
+export { useGenerateRandomTemp } from "./model/useGenerateRandomTemp";
