+++ src/features/random-temp/index.js
@@ -1,0 +1,1 @@
+export { default as RandomTemp } from "./ui/RandomTemp";
