--- src/hooks/useModal.jsx
@@ -1,15 +1,0 @@
-import { useSubmitActions } from "../store/uiState";
-
-export function useModal() {
-  const { setResult } = useSubmitActions();
-
-  const close = () => {
-    setResult(false);
-  };
-
-  const confirm = () => {
-    setResult(false);
-  };
-
-  return { close, confirm };
-}
