--- src/hooks/useNavigator.jsx
@@ -1,13 +1,0 @@
-import { useNavigate, useOutletContext } from "react-router-dom";
-
-export function useGoHome() {
-  const { handleHidden } = useOutletContext();
-  const navigate = useNavigate();
-
-  const handleClick = () => {
-    handleHidden();
-    navigate("/");
-  };
-
-  return { handleClick };
-}
