+++ src/pages/random-temp/index.js
@@ -1,0 +1,2 @@
+export { default as RandomFoodTemp } from "./food/RandomFoodTemp";
+export { default as RandomMachineTemp } from "./machine/RandomMachineTemp";
