+++ src/shared/lib/hooks/index.js
@@ -1,2 +1,3 @@
 export { useModal } from "./use-modal";
+export { useModalWithTemp } from "./use-modal";
 export { useNavigateToMain } from "./use-navigate-to-main";
