+++ src/shared/lib/hooks/use-modal.js
@@ -1,4 +1,5 @@
 import { useState } from "react";
+import { useSubmitActions } from "../../../entities/ui-state";
 
 export function useModal(onCloseCallback) {
   const [openModal, setOpenModal] = useState(false);
@@ -14,4 +15,19 @@ export function useModal(onCloseCallback) {
   };
 
   return { openModal, handleModalOpen, handleModalClose };
+}
+
+export function useModalWithTemp(handleRetrySubmit) {
+  const { setResult } = useSubmitActions();
+
+  const handleModalClose = () => {
+    setResult(false);
+  };
+
+  const handleRetry = () => {
+    handleRetrySubmit();
+    setResult(false);
+  };
+
+  return { handleModalClose, handleRetry };
 }
