+++ src/widgets/random-temp-form/index.js
@@ -1,0 +1,1 @@
+export { default as RandomTempForm } from "./ui/RandomTempForm";
